backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.3",
    "google-adk>=1.28.1",
    "google-genai>=1.70.0",
    "google-generativeai>=0.8.6",
    "greenlet>=3.3.2",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.43.0",
]
